Create the class ImageHandler and refactor
