+++ core/src/View/View.java
@@ -2,11 +2,8 @@ package View;
 
 import Model.*;
 import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.assets.loaders.SoundLoader;
 import com.badlogic.gdx.graphics.OrthographicCamera;
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
-import com.badlogic.gdx.graphics.g2d.TextureRegion;
 
 import com.badlogic.gdx.graphics.GL20;
 import com.badlogic.gdx.maps.tiled.TiledMap;
@@ -15,17 +12,14 @@ import com.badlogic.gdx.maps.tiled.TmxMapLoader;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
 
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.HashSet;
 import java.util.Set;
-import java.util.concurrent.*;
 
 
 public class View {
     private Model model;
     private float timeSincePlayerWalkFrameChanged = 0f;
     private int currentPlayerWalkFrame = 0;
-    private HashMap<String, TextureRegion[][]> allEntityImages = new HashMap<String, TextureRegion[][]>();
     private OrthographicCamera camera;
     private SpriteBatch batch;
     private static final int SCREEN_WIDTH = 800;
@@ -34,6 +28,7 @@ public class View {
     private TiledMap tiledMap;
     private TiledMapRenderer tiledMapRenderer;
     private ImageLoader imageLoader = new ImageLoader();
+    private ImageHandler imageHandler = new ImageHandler();
     private Sound soundLoader = new Sound();
     private int i = 0;
 
@@ -44,7 +39,7 @@ public class View {
     }
 
     public void initialize() {
-        loadEntityImages();
+        imageHandler.loadEntityImages();
 
         camera = new OrthographicCamera();
         camera.setToOrtho(false, SCREEN_WIDTH, SCREEN_HEIGHT);
@@ -68,7 +63,7 @@ public class View {
         batch.setProjectionMatrix(camera.combined);
         batch.begin();
         updatePlayerWalkFrame();
-        batch.draw(getEntityImage("Player", model.getPlayerDirection(), currentPlayerWalkFrame), model.getPlayer().getX(), model.getPlayer().getY());
+        batch.draw(imageHandler.getEntityImage("Player", model.getPlayerDirection(), currentPlayerWalkFrame), model.getPlayer().getX(), model.getPlayer().getY());
         drawAllEntities(model);
         batch.end();
 
@@ -102,31 +97,6 @@ public class View {
     }
 
     /**
-     * Loads a sprite sheet for the specified entity, splits it into multiple images
-     * and returns a 2D array which contains them.
-     * @param entityName the name of the entity
-     * @return a 2D array with texture regions of the sprite sheet of the specified entity
-     */
-    private TextureRegion[][] getEntityImages(String entityName) {
-        Texture spriteSheet = new Texture(Gdx.files.internal("entities/" + entityName + "SpriteSheet.png"));
-        int nColumnsPlayerWalkSheet = 4;
-        int nRowsPlayerWalkSheet = 4;
-        // Splits the sprite sheet into multiple images
-        return TextureRegion.split(spriteSheet,
-                spriteSheet.getWidth() / nColumnsPlayerWalkSheet,
-                spriteSheet.getHeight() / nRowsPlayerWalkSheet);
-    }
-
-    /**
-     * Loads the entity images and puts them in the HashMap allEntityImages.
-     */
-    private void loadEntityImages() {
-        allEntityImages.put("Player", getEntityImages("Player"));
-        allEntityImages.put("Mouse", getEntityImages("Mouse"));
-        allEntityImages.put("Cyclops", getEntityImages("Cyclops"));
-    }
-
-    /**
      * Changes the player walk frame to the next frame in the walk animation after a certain time interval if the player is moving.
      * Otherwise, the frame is set to the first frame of the walk animation.
      */
@@ -142,29 +112,6 @@ public class View {
         else {
             currentPlayerWalkFrame = 0;
             timeSincePlayerWalkFrameChanged = 0f;
-        }
-    }
-
-    /**
-     * Gets the entity image based on the direction of the entity and the current frame in the walk animation.
-     * @param entityName the name of the entity
-     * @param currentFrame the current frame in the entity walk animation
-     * @param direction the direction that the entity is facing
-     * @return the texture region which matches the state of the entity
-     */
-    private TextureRegion getEntityImage(String entityName, Direction direction, int currentFrame) {
-        TextureRegion[][] entityImages = allEntityImages.get(entityName);
-        switch(direction) {
-            case UP:
-                return entityImages[currentFrame][1];
-            case DOWN:
-                return entityImages[currentFrame][0];
-            case LEFT:
-                return entityImages[currentFrame][2];
-            case RIGHT:
-                return entityImages[currentFrame][3];
-            default:
-                throw new AssertionError();
         }
     }
 }
